+++ src/CoachConnect.API/Controllers/CoachesController.cs
@@ -1,4 +1,5 @@
-﻿using CoachConnect.BusinessLayer.DTOs;
+﻿using CoachConnect.BusinessLayer.DTOs.Coach;
+using CoachConnect.BusinessLayer.DTOs.Users;
 using CoachConnect.BusinessLayer.Services;
 using CoachConnect.BusinessLayer.Services.Interfaces;
 using CoachConnect.Shared.Helpers;
