+++ src/CoachConnect.API/appsettings.json
@@ -2,12 +2,12 @@
   "ConnectionStrings": {
     "DefaultConnection": "Server=localhost;Database=coach_connect;User ID=coach-app;Password=coach123;" // husk sette inn Port for evt db docker container
   },
-    
+
   "Jwt": {
     "Key": "ThisismySecretKeyDoNotStoreHereForGodsSake",
     "Issuer": "CoachConnect.com"
   },
-  
+
   "Serilog": {
     "Using": [ "Serilog.Sinks.Console", "Serilog.Sinks.File" ],
     "MinimumLevel": {
@@ -33,7 +33,7 @@
       {
         "Name": "MySQL",
         "Args": {
-          "connectionString": "Server=localhost;Database=coach_connect;User ID=coach-app;Password=coach123;", 
+          "connectionString": "Server=localhost;Database=coach_connect;User ID=coach-app;Password=coach123;",
           "tableName": "Logs",
           "autoCreateSqlTable": true
         }
