+++ src/CoachConnect.BusinessLayer/DTOs/Coach/CoachDTO.cs
@@ -1,11 +1,11 @@
 ﻿using CoachConnect.DataAccess.Entities;
 
-namespace CoachConnect.BusinessLayer.DTOs;
+namespace CoachConnect.BusinessLayer.DTOs.Coach;
 
 public record CoachDTO(
     string FirstName,
-    string LastName,    
-    string PhoneNumber,    
+    string LastName,
+    string PhoneNumber,
     string Email,
     CoachId Id,
     IEnumerable<TeamDTO> Teams);
