+++ src/CoachConnect.BusinessLayer/DTOs/Coach/CoachRegistrationDTO.cs
@@ -1,11 +1,11 @@
 ﻿using CoachConnect.DataAccess.Entities;
 
-namespace CoachConnect.BusinessLayer.DTOs;
+namespace CoachConnect.BusinessLayer.DTOs.Coach;
 
 public record CoachRegistrationDTO(
     string FirstName,
     string LastName,
     string PhoneNumber,
     string Password,
-    string Email    
+    string Email
     );
