+++ src/CoachConnect.BusinessLayer/DTOs/User/UserCoachUpdateDTO.cs
@@ -1,8 +1,7 @@
-﻿namespace CoachConnect.BusinessLayer.DTOs;
+﻿namespace CoachConnect.BusinessLayer.DTOs.Users;
 
 public record UserCoachUpdateDTO(
     string FirstName,
     string LastName,
     string PhoneNumber,
     string Email);
-
