+++ src/CoachConnect.BusinessLayer/DTOs/User/UserRegistrationDTO.cs
@@ -1,6 +1,6 @@
-﻿namespace CoachConnect.BusinessLayer.DTOs;
+﻿namespace CoachConnect.BusinessLayer.DTOs.Users;
 
-public record UserRegistrationDTO( 
+public record UserRegistrationDTO(
     string FirstName,
     string LastName,
     string PhoneNumber,
