+++ src/CoachConnect.BusinessLayer/Mappers/CoachMappers/CoachMapper.cs
@@ -1,8 +1,9 @@
 ﻿using CoachConnect.BusinessLayer.DTOs;
+using CoachConnect.BusinessLayer.DTOs.Coach;
 using CoachConnect.BusinessLayer.Mappers.Interfaces;
 using CoachConnect.DataAccess.Entities;
 
-namespace CoachConnect.BusinessLayer.Mappers;
+namespace CoachConnect.BusinessLayer.Mappers.Coaches;
 
 public class CoachMapper : IMapper<Coach, CoachDTO>
 {
