+++ src/CoachConnect.BusinessLayer/Mappers/GameMapper.cs
@@ -11,21 +11,23 @@ public class GameMapper : IMapper<Game, GameDTO>
         return new GameDTO(
             entity.Location,
             entity.OpponentName,
-            entity.GameTime,
-            entity.Id
+            entity.GameTime.ToString("f"),
+            entity.Id.gameId
             );
     }
 
     public Game MapToEntity(GameDTO dto)
     {
-        var dtnow = DateTime.Now;
-        return new Game
-        {
-            Location = dto.Location,
-            OpponentName = dto.OpponentName,
-            GameTime = dto.GameTime,
-            Created = dtnow,
-            Updated = dtnow
-        };
+        throw new NotImplementedException();
+
+        //var dtnow = DateTime.Now;
+        //return new Game
+        //{
+        //    Location = dto.Location,
+        //    OpponentName = dto.OpponentName,
+        //    GameTime = dto.GameTime,
+        //    Created = dtnow,
+        //    Updated = dtnow
+        //};
     }
 }
