+++ src/CoachConnect.BusinessLayer/Mappers/UserMappers/UserMapper.cs
@@ -1,8 +1,9 @@
 ﻿using CoachConnect.BusinessLayer.DTOs;
+using CoachConnect.BusinessLayer.DTOs.Users;
 using CoachConnect.BusinessLayer.Mappers.Interfaces;
 using CoachConnect.DataAccess.Entities;
 
-namespace CoachConnect.BusinessLayer.Mappers;
+namespace CoachConnect.BusinessLayer.Mappers.UserMappers;
 
 public class UserMapper : IMapper<User, UserDTO>
 {
@@ -11,23 +12,22 @@ public class UserMapper : IMapper<User, UserDTO>
         return new UserDTO()
         {
             FirstName = entity.FirstName,
-            LastName = entity.LastName,            
-            PhoneNumber = entity.PhoneNumber, 
-            Email = entity.Email, 
+            LastName = entity.LastName,
+            PhoneNumber = entity.PhoneNumber,
+            Email = entity.Email,
             Id = entity.Id,
             Players = new List<PlayerDTO>(),
         };
     }
 
-    public User MapToEntity(UserDTO dto) 
+    public User MapToEntity(UserDTO dto)
     {
-        // var dtNow = DateTime.Now;
         return new User()
-        {          
-            FirstName = dto.FirstName, 
+        {
+            FirstName = dto.FirstName,
             LastName = dto.LastName,
             PhoneNumber = dto.PhoneNumber,
-            Email = dto.Email,            
+            Email = dto.Email,
         };
     }
 }
