+++ src/CoachConnect.BusinessLayer/Services/Interfaces/ICoachService.cs
@@ -1,4 +1,5 @@
-﻿using CoachConnect.BusinessLayer.DTOs;
+﻿using CoachConnect.BusinessLayer.DTOs.Coach;
+using CoachConnect.BusinessLayer.DTOs.Users;
 using CoachConnect.DataAccess.Entities;
 using CoachConnect.Shared.Helpers;
 using System;
